Bypassing FPDF Issue
FPDF issue still not resolved and have wasted a lot of time trying to fix this. Plan to comment it out in the script and carry on, hoping to come back if it can be resolved.

Copied practical 2 Cartopy and modified and added data so that Cyprus is mapped to a png.